[ 1.0.3 ] add robby
Users cannot make their matchmaking freely; Server just assigns a room_id to Users.

+++ The Tree House Server/The Tree House Server/GameObject.h
@@ -19,7 +19,7 @@ public:
 	Transform GetTransform();
 
 public:
-	unsigned short m_id;
+	unsigned int m_id;
 
 	iVector3   m_position;
 	sRotation3 m_rotation;
